Added codebase for Peers Controller
